PENV-40: move proxy dir from application to applacationbase
