+++ .gitignore
@@ -52,5 +52,4 @@ network_cache
 /docs/build
 
 # contracts generated files
-application/contract
-application/proxy
+applicationbase/proxy
