+++ application/functest/logicrunner_test.go
@@ -85,7 +85,7 @@ import (
 	"errors"
 
 	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	"github.com/insolar/insolar/application/proxy/two"
+	"github.com/insolar/insolar/applicationbase/proxy/two"
 	"github.com/insolar/insolar/insolar"
 )
 
@@ -265,7 +265,7 @@ package main
 import (
 "github.com/insolar/insolar/insolar"
 "github.com/insolar/insolar/logicrunner/builtin/foundation"
-"github.com/insolar/insolar/application/proxy/test_saga_simple_contract"
+"github.com/insolar/insolar/applicationbase/proxy/test_saga_simple_contract"
 )
 
 type TestSagaSimpleCallContract struct {
@@ -353,7 +353,7 @@ package main
 import (
 "github.com/insolar/insolar/insolar"
 "github.com/insolar/insolar/logicrunner/builtin/foundation"
-"github.com/insolar/insolar/application/proxy/test_saga_call_from_accept_method"
+"github.com/insolar/insolar/applicationbase/proxy/test_saga_call_from_accept_method"
 )
 
 type TestSagaCallFromAcceptMethodContract struct {
@@ -466,7 +466,7 @@ package main
 import (
 "github.com/insolar/insolar/insolar"
 "github.com/insolar/insolar/logicrunner/builtin/foundation"
-"github.com/insolar/insolar/application/proxy/test_saga_multiple_calls"
+"github.com/insolar/insolar/applicationbase/proxy/test_saga_multiple_calls"
 )
 
 type TestSagaMultipleCallsContract struct {
@@ -564,7 +564,7 @@ package main
 import (
 "github.com/insolar/insolar/insolar"
 "github.com/insolar/insolar/logicrunner/builtin/foundation"
-"github.com/insolar/insolar/application/proxy/test_saga_magic_flag_contract_two"
+"github.com/insolar/insolar/applicationbase/proxy/test_saga_magic_flag_contract_two"
 )
 
 type SagaMagicFlagOne struct {
@@ -666,7 +666,7 @@ package main
 
 import (
 "github.com/insolar/insolar/logicrunner/builtin/foundation"
-"github.com/insolar/insolar/application/proxy/test_saga_self_contract"
+"github.com/insolar/insolar/applicationbase/proxy/test_saga_self_contract"
 )
 
 type TestSagaSelfCallContract struct {
@@ -794,7 +794,7 @@ package main
 
 import (
 	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	two "github.com/insolar/insolar/application/proxy/error_interface_two"
+	two "github.com/insolar/insolar/applicationbase/proxy/error_interface_two"
 )
 
 type One struct {
@@ -872,7 +872,7 @@ package main
 
 import (
 	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	two "github.com/insolar/insolar/application/proxy/nil_result_two"
+	two "github.com/insolar/insolar/applicationbase/proxy/nil_result_two"
 )
 
 type One struct {
@@ -931,7 +931,7 @@ package main
 
 import (
 	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	two "github.com/insolar/insolar/application/proxy/constructor_return_nil_two"
+	two "github.com/insolar/insolar/applicationbase/proxy/constructor_return_nil_two"
 )
 
 type One struct {
@@ -987,7 +987,7 @@ package main
 
 import (
 	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	two "github.com/insolar/insolar/application/proxy/constructor_return_error_two"
+	two "github.com/insolar/insolar/applicationbase/proxy/constructor_return_error_two"
 )
 
 type One struct {
@@ -1041,7 +1041,7 @@ func TestGetParent(t *testing.T) {
 
  import (
 	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	two "github.com/insolar/insolar/application/proxy/get_parent_two"
+	two "github.com/insolar/insolar/applicationbase/proxy/get_parent_two"
  )
 
  type One struct {
@@ -1105,7 +1105,7 @@ package main
 
 import (
 	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	two "github.com/insolar/insolar/application/proxy/get_remote_data_two"
+	two "github.com/insolar/insolar/applicationbase/proxy/get_remote_data_two"
 )
 
 type One struct {
@@ -1158,7 +1158,7 @@ package main
 
 import (
 	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	two "github.com/insolar/insolar/application/proxy/immutable_annotation_two"
+	two "github.com/insolar/insolar/applicationbase/proxy/immutable_annotation_two"
 )
 
 type One struct {
@@ -1195,7 +1195,7 @@ package main
 
 import (
 	"github.com/insolar/insolar/logicrunner/builtin/foundation"
-	three "github.com/insolar/insolar/application/proxy/immutable_annotation_three"
+	three "github.com/insolar/insolar/applicationbase/proxy/immutable_annotation_three"
 )
 
 type Two struct {
