+++ application/functest/pressure_test.go
@@ -137,7 +137,7 @@ package main
 
 import "github.com/insolar/insolar/insolar"
 import "github.com/insolar/insolar/logicrunner/builtin/foundation"
-import "github.com/insolar/insolar/application/proxy/testCoinPassing"
+import "github.com/insolar/insolar/applicationbase/proxy/testCoinPassing"
 import "errors"
 
 type One struct {
